ground-control/requirements.txt
aiohttp>=3.9.0
aiofiles>=23.0.0
pyyaml>=6.0
watchdog>=3.0.0
